Thanks for writing this shard. It's really well done!

The only thing that I'm stumbling over is that I can't have multiple label sets for the same metric. For example, it doesn't seem like I can emit the metrics `http_request_count{status_class="2x"}` and `http_request_count{status_class="5xx"}` using the same registry. 

This PR adds support for that by giving each metric a data store instead of a single scalar value. Method calls on the metric delegate to the data store, which tracks all of the values for all of the label sets used.

There are some things to improve in this PR, so it's not done yet. For example, I implemented histograms and summaries in terms of counters (one for count, one for sum, and one for each bucket) because I didn't want to have to duplicate all of that logic, however that means it N+1 mutex calls (one for the histogram and one for each counter … for _each label set_), which has performance implications. I'm considering adding a `protected def Counter#inc!` method that bypasses the mutex to mitigate that.

I wrote a quick benchmark to see what the performance implications would be. As expected, it had the largest impact on histograms, which were already the slowest and are known to be heavy in Prometheus simply due to the quantity of time series they require.

<details><summary>Benchmark code</summary>

```crystal
require "benchmark"
require "../src/prometheus"

Benchmark.ips do |x|
  counter = Prometheus.counter("http_requests_total", "Total HTTP Requests")
  gauge = Prometheus.gauge "cpu_usage", "CPU usage percentage"
  histogram = Prometheus.histogram "http_response_time", "HTTP response times in seconds",
    buckets: [0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 2, 3, 5, 8, 13, 21, 30] of Float64
  summary = Prometheus.summary "http_response_errors", "HTTP Response Errors"

  x.report "counter" { counter.inc }
  x.report "gauge" { gauge.set 42 }
  x.report "histogram" { histogram.observe 0.25 }
  x.report "summary" { summary.observe 1 }
end
```

</details>

### Before

```
➜  crystal-prometheus-client git:(main) crr benchmark/metrics.cr
  counter 135.30M (  7.39ns) (± 5.64%)  0.0B/op   1.05× slower
    gauge 141.67M (  7.06ns) (± 3.89%)  0.0B/op        fastest
histogram   8.08M (123.75ns) (± 6.17%)  0.0B/op  17.53× slower
  summary 140.29M (  7.13ns) (± 3.20%)  0.0B/op   1.01× slower
```

### After

```
➜  crystal-prometheus-client git:(allow-multiple-labelsets-per-metric) crr benchmark/metrics.cr
  counter  85.96M ( 11.63ns) (± 2.59%)  0.0B/op   1.19× slower
    gauge 102.30M (  9.77ns) (± 3.82%)  0.0B/op        fastest
histogram   3.16M (316.75ns) (± 0.41%)  0.0B/op  32.40× slower
  summary  33.54M ( 29.81ns) (± 2.55%)  0.0B/op   3.05× slower
```

I don't think much can be done to improve the `Counter` and `Gauge` types with this PR. Adding 2ns of overhead per hash dereference (there's more of a difference for `Counter` than `Gauge` because it does a read _and_ a write via the data store hash whereas `Gauge` just does a write) is unavoidable. Removing the nested/repeated mutex usage should improve the histogram and summary performance dramatically, though. In other benchmarks I've run, I've found that each mutex lock is around 10ns. I added an unreasonable number of buckets to the histogram (the old rule of thumb for Prometheus was a cardinality of 10-12 and this histogram has 22 without any labels) as a pathological case to see what the performance hit would be there.