@pfeifferj 2025-05-03
hiya! thank you for the thoughtful PR, I really appreciate it! let me know it's ready for review :)