@jgaskins 2025-05-03
TL;DR: This is ready for review and, while it's still slower than on `main`, it's 33% faster than it was last night.

---

After https://github.com/pfeifferj/crystal-prometheus-client/pull/1/commits/56319126bce133f436d2ce5bd94e2238125909df I was able to reduce histogram and summary CPU costs by about 1/3. This is the before and after for that commit specifically:

```
➜  crystal-prometheus-client git:(allow-multiple-labelsets-per-metric) crr benchmark/metrics.cr
  counter  83.90M ( 11.92ns) (± 0.76%)  0.0B/op   1.19× slower
    gauge  99.95M ( 10.00ns) (± 1.28%)  0.0B/op        fastest
histogram   3.06M (326.73ns) (± 1.21%)  0.0B/op  32.66× slower
  summary  33.11M ( 30.20ns) (± 2.98%)  0.0B/op   3.02× slower
➜  crystal-prometheus-client git:(allow-multiple-labelsets-per-metric) crr benchmark/metrics.cr
  counter  83.87M ( 11.92ns) (± 2.57%)  0.0B/op   1.21× slower
    gauge 101.31M (  9.87ns) (± 0.99%)  0.0B/op        fastest
histogram   4.35M (229.78ns) (± 0.70%)  0.0B/op  23.28× slower
  summary  52.39M ( 19.09ns) (± 0.44%)  0.0B/op   1.93× slower
```

It's still a fair bit slower than the current `main` branch, so I ran this code through macOS Instruments.app:

```crystal
require "../src/prometheus"

histogram = Prometheus.histogram "foo", "",
  buckets: [0.005, 0.01, 0.025, 0.05, 0.1, 0.25, 0.5, 1, 2.5, 5, 10] of Float64

loop do
  histogram.observe 0.15
end
```

What I found was that 52% of CPU time after the latest commit was spent in `Hash(String, String)#[]=`, which aligns pretty closely with my benchmarks — growth from 123ns (for the single buckets hash as implemented on `main`) to 229ns behind a second hash (as of 56319126bce133f436d2ce5bd94e2238125909df) means the gain is 46% of the total time spent, so nearly all of that gain was due specifically to a second layer of hash iterations and assignments.