@pfeifferj 2025-05-04
I tested things with:
crystal `1.16.1`
ameba `1.4.3`

and they worked fine. you might want to remove the `version: ~> 1.4.0` version constraint for ameba because that caused errors for me.

looking at the code changes now